likwid collector: explicitly cast input to HPMaddthread helper function to int. See #164

+++ collectors/likwidMetric.go
@@ -337,7 +337,7 @@ func (m *LikwidCollector) Init(config json.RawMessage) error {
 		for _, c := range m.cpulist {
 			m.measureThread.Call(
 				func() {
-					retCode := C._HPMaddThread(c)
+					retCode := C._HPMaddThread(C.int(c))
 					if retCode != 0 {
 						err := fmt.Errorf("C.HPMaddThread(%v) failed with return code %v", c, retCode)
 						cclog.ComponentError(m.name, err.Error())
